requirements.txt
requests
google-genai
newspaper3k
lxml_html_clean
